+++ src/mesh/RadioInterface.cpp
@@ -1489,7 +1489,7 @@ size_t RadioInterface::beginSending(meshtastic_MeshPacket *p)
     }
 
     if (static_cast<size_t>(p->encrypted.size) > sizeof(radioBuffer.payload)) {
-        LOG_ERROR("Packet payload size %zu exceeds radioBuffer capacity %zu", static_cast<size_t>(p->encrypted.size), sizeof(radioBuffer.payload));
+        LOG_ERROR("Packet payload size %u exceeds radioBuffer capacity %u", static_cast<unsigned>(p->encrypted.size), static_cast<unsigned>(sizeof(radioBuffer.payload)));
         packetPool.release(p);
         return 0;
     }
